backend/requirements.txt
python_jose[cryptography]==3.2.0
passlib[bcrypt]==1.7.4
asyncpg==0.21.0
rejson==0.5.4
pandas==1.1.2
pydantic==1.6.1
tortoise_orm==0.16.16
fastapi==0.61.1
uvicorn
python-multipart
databases==0.4.0
sqlalchemy==1.3.20
fastapi-users[tortoise-orm]==3.0.6